Add(products,profiles): view security @login_req
Add superuser and login checks on product and profile edit/delete views

+++ products/views.py
@@ -4,6 +4,7 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
 # Internal:
 from .models import Product, Category
 from .forms import ProductForm
@@ -80,12 +81,17 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     View for adding products to the database
         Arguments: request (object): The Http request
         Returns: Render add product page
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry this feature is for store owners only.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -107,6 +113,7 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     View for editing product details
@@ -116,6 +123,10 @@ def edit_product(request, product_id):
         Returns:
             Render the edit product page
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry this feature is for store owners only.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -141,6 +152,7 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     View for deleting product details
@@ -150,6 +162,10 @@ def delete_product(request, product_id):
         Returns:
             Redirect to products page
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry this feature is for store owners only.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product removed successfully')
